+++ src/interact/cli/cli.ts
@@ -7,7 +7,7 @@ import {config} from './commands/config.js'
 import {favicon} from './commands/favicon.js'
 import {schema} from './commands/schema.js'
 import type {LogLevel} from './shared/vite.config.js'
-import {version} from '../../../package.json';
+import packageFile from '../../../package.json' with {type: "json"};
 
 const cli = cac('interact')
 
@@ -98,7 +98,7 @@ cli
     })
 
 cli.help()
-cli.version(version)
+cli.version(packageFile.version)
 
 export async function run(argv: string[] = process.argv): Promise<void> {
     try {