+++ src/resources/components/partials/Aside.tsx
@@ -1,6 +1,8 @@
-import type {Frontmatter, Page, LayoutProps} from "@combostrap/interact/types";
+import type {Frontmatter, LayoutProps, Page} from "@combostrap/interact/types";
 import React from "react";
-import Tree, {type TreeNode} from "@/components/interact/Tree.tsx";
+import Tree from "@/components/interact/Tree.tsx";
+import {getInteractConfig} from "../../../interact/config/interactConfig.ts";
+import {getPagesTree} from "../../rsc/server/handler.tsx";
 
 const pages = import.meta.glob<Page<Frontmatter>>(
     "./pages/**/*.ts",
@@ -14,10 +16,11 @@ function toRoute(path: string) {
         .replace(/\.ts$/, "")
 }
 
+
 export const nav = Object.entries(pages)
     .map(([path, mod]) => ({
-        route: toRoute(path),
-        title: mod.frontmatter?.title ?? "Untitled",
+        path: toRoute(path),
+        name: mod.frontmatter?.title ?? "Untitled",
         order: mod.frontmatter?.order ?? 0,
         group: mod.frontmatter?.group
     }))
@@ -27,43 +30,8 @@ export type AsideProps = React.HTMLAttributes<HTMLElement> & LayoutProps
 
 // @ts-ignore -- exported
 export default function Aside({page, context, ...props}: AsideProps) {
-
-    const data: TreeNode[] = [
-        {
-            name: "src",
-            path: "/src",
-            type: "folder",
-            children: [
-                {
-                    name: "components",
-                    path: "/src/components",
-                    type: "folder",
-                    children: [
-                        {
-                            name: "Button.tsx",
-                            path: "/src/components/Button.tsx",
-                            type: "file",
-                        },
-                        {
-                            name: "Sidebar.tsx",
-                            path: "/src/components/Sidebar.tsx",
-                            type: "file",
-                        },
-                    ],
-                },
-                {
-                    name: "app.tsx",
-                    path: "/src/app.tsx",
-                    type: "file",
-                },
-            ],
-        },
-        {
-            name: "package.json",
-            path: "/package.json",
-            type: "file",
-        },
-    ]
+    let interactConfig = getInteractConfig();
+    let data = getPagesTree(interactConfig.paths.pagesDirectory).children ?? []
 
     return <Tree data={data} {...props} />