+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+* Added [Tree component](src/resources/components/interact/Tree.tsx)
+* Change [Aside partial](src/resources/components/partials) to show the tree component
+
 ## [0.1.1] - 2026-06-18
 
 * First version